refactor: split party setup dialogs
